Powerups, Bullet rotation, Enemy Spawning, Power-up UI,
Finished the first implementation of the power-ups. Add a simple UI for the power-ups so you could see what is active. Made it so there is a max of 2 per category. Added simple vfx on the power-ups so you know when it is active/triggered. Made it so the bullet goes towards where ever the player/enemy is facing.