Crit VFX, Health Display
Added a crit flash effect similar to the damage effect. So instead of flashing white it will flash yellow when there's a crit since I added it in the Entity class meaning the player could be critted as well in the future. Added the health text to game manager. Deleted Explosive Bullets Powerup since we will not be using that. Re-purposed Explosive Bullets prefab for Vampiric Rounds.